Refatoração de código
