+++ monster.h
@@ -1,3 +1,5 @@
+// Biblioteca para as funções relacionadas aos monstros
+
 #include <stdlib.h>
 #include <math.h>
 #include <GL/glut.h>
